Address PR comments  * Move parameters to Parameter folder  * Remove the spurious SecurityTokenKindParameter file
